requirements.txt
# =============================================================================
# Cloud Node Anomaly Detection - Python Dependencies
# =============================================================================

# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Deep Learning (TensorFlow/Keras)
tensorflow>=2.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Explainability
shap>=0.41.0
lime>=0.2.0

# Dashboard
streamlit>=1.20.0

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0

